test: move ldapjs Client mock to mock folder
